Separated the Movie/TV show activity from the MainActivity into a seperqte Activity
