+++ app/src/main/AndroidManifest.xml
@@ -10,14 +10,23 @@
         android:label="@string/app_name"
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
-        <activity android:name=".MainActivity"
-            android:configChanges="keyboardHidden|orientation|screenSize">
+        <activity android:name=".MainActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
 
                 <category android:name="android.intent.category.LAUNCHER" />
+
             </intent-filter>
+
         </activity>
+
+        <activity
+            android:name=".IndividualPage"
+            android:configChanges="keyboardHidden|orientation|screenSize">
+
+        </activity>
+
+
     </application>
 
 </manifest>
