+++ app/src/main/java/com/example/com/reelreviews/Config.java
@@ -6,7 +6,7 @@ package com.example.com.reelreviews;
 
 public final class Config {
 
-
     public static final String API_KEY = "AIzaSyC8frZQ_xjSeSqeBJIpkPnnKkljmIyvPTw";
     public static String video_String = "_MC3XuMvsDI";
+
 }
