+++ app/src/main/res/layout/activity_main.xml
@@ -1,138 +1,15 @@
 <?xml version="1.0" encoding="utf-8"?>
-<ScrollView xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:id="@+id/activity_main"
+<RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
     android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    tools:context="com.example.com.reelreviews.MainActivity">
+    android:layout_height="match_parent">
 
-    <LinearLayout
-        android:layout_width="match_parent"
+    <Button
+        android:id="@+id/button"
+        android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:fitsSystemWindows="true"
-        android:orientation="vertical">
-
-        <include layout="@layout/toolbar_main" />
-
-        <RelativeLayout
-            android:layout_width="match_parent"
-            android:layout_height="match_parent"
-            android:layout_marginLeft="@dimen/activity_horizontal_margin"
-            android:layout_marginRight="@dimen/activity_horizontal_margin"
-            android:layout_marginTop="@dimen/activity_vertical_margin">
-
-            <TextView
-                android:id="@+id/title"
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content"
-                android:text="Inside Out (2015)"
-                android:textAlignment="center"
-                android:textSize="20sp" />
-
-            <ImageView
-                android:id="@+id/thumbnail"
-                android:layout_width="130dp"
-                android:layout_height="195dp"
-                android:layout_below="@id/title"
-                android:layout_marginTop="@dimen/activity_vertical_margin"
-                android:adjustViewBounds="true"
-                android:cropToPadding="false"
-                android:scaleType="fitXY"
-                android:src="@drawable/place_holder_insideout" />
-
-            <LinearLayout
-                android:layout_width="wrap_content"
-                android:layout_height="195dp"
-                android:layout_below="@id/title"
-                android:layout_marginLeft="@dimen/activity_horizontal_margin"
-                android:layout_marginStart="@dimen/activity_horizontal_margin"
-                android:layout_marginTop="@dimen/activity_vertical_margin"
-                android:layout_toEndOf="@id/thumbnail"
-                android:layout_toRightOf="@id/thumbnail"
-                android:orientation="vertical">
-
-                <TextView
-                    android:id="@+id/reelreview"
-                    android:layout_width="wrap_content"
-                    android:layout_height="wrap_content"
-                    android:layout_weight="1"
-                    android:text="Reel Rating - 90"
-                    android:textSize="16sp" />
-
-                <TextView
-                    android:id="@+id/rottenTomatoes"
-                    android:layout_width="wrap_content"
-                    android:layout_height="wrap_content"
-                    android:layout_marginTop="20dp"
-                    android:layout_weight="1"
-                    android:text="Rotten Tomatoes - 60"
-                    android:textSize="16sp" />
-
-                <TextView
-                    android:id="@+id/metacritic"
-                    android:layout_width="wrap_content"
-                    android:layout_height="wrap_content"
-                    android:layout_marginTop="20dp"
-                    android:layout_weight="1"
-                    android:text="Metacritic - 30"
-                    android:textSize="16sp" />
-
-                <TextView
-                    android:id="@+id/imdb"
-                    android:layout_width="wrap_content"
-                    android:layout_height="wrap_content"
-                    android:layout_marginTop="20dp"
-                    android:layout_weight="1"
-                    android:text="IMDB - 60"
-                    android:textSize="16sp" />
-
-
-            </LinearLayout>
-
-
-        </RelativeLayout>
-
-        <TextView
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:layout_marginLeft="@dimen/activity_horizontal_margin"
-            android:layout_marginTop="@dimen/activity_vertical_margin"
-            android:text="Synopsis"
-            android:textSize="20sp" />
-
-        <TextView
-            android:id="@+id/synopsis"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:layout_marginLeft="@dimen/activity_horizontal_margin"
-            android:layout_marginTop="@dimen/activity_vertical_margin"
-            android:text="@string/synopsisContent"
-            android:textSize="14sp" />
-
-        <fragment
-            android:id="@+id/youtubepLAYER_fragment"
-            android:name="com.google.android.youtube.player.YouTubePlayerSupportFragment"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:layout_marginTop="@dimen/activity_vertical_margin" />
-
-        <TextView
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:layout_marginLeft="@dimen/activity_horizontal_margin"
-            android:layout_marginTop="@dimen/activity_vertical_margin"
-            android:text="Cast"
-            android:textSize="20dp" />
-
-        <TextView
-            tools:text="Joy\nSally\nRandy\nChuck"
-            android:id="@+id/cast"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:layout_marginLeft="@dimen/activity_horizontal_margin"
-            android:layout_marginTop="@dimen/activity_vertical_margin"
-            android:textSize="14sp" />
-
-    </LinearLayout>
-
-</ScrollView>
+        android:layout_alignParentTop="true"
+        android:layout_centerHorizontal="true"
+        android:layout_marginTop="81dp"
+        android:onClick="openPage"
+        android:text="Open Page" />
+</RelativeLayout>
